books routes protected

+++ backend/controllers/bookController.js
@@ -3,6 +3,7 @@
 const Book = require('../models/book');
 const mongoose = require('mongoose');
 const { isValidObjectId } = mongoose;
+const User = require('../models/user');
 
 // Function to fetch all books
 exports.getAllBooks = async (req, res) => {
@@ -33,11 +34,33 @@ exports.getABook = async (req, res) => {
 }
 
 
+
+
+
 exports.deleteBook = async (req, res) => {
   try {
     if (!isValidObjectId(req.params.id)) {
       return res.status(400).json({ error: 'Invalid ID' });
     }
+
+    // find the book by ID
+    const book = await Book.findById(req.params.id);
+    if (!book) {
+      return res.status(400).json({ error: 'Book not found'});
+    }
+
+    // 
+    const user = await User.findById(req.user.id);
+    if (!book.user) {
+      return res.status(400).json({ error: 'User not found' });
+    }
+
+    // Verify that the user who want to delete the book is the owner
+    if (book.user.toString() !== user.id) {
+      return res.status(401).json({ error: 'Only the owner can delete this book' });;
+    }
+
+    // Delete the book
     const bookToDelete = await Book.findByIdAndDelete(req.params.id);
     if (!bookToDelete) {
       return res.status(400).json({ error: 'Book not found' });
@@ -50,9 +73,11 @@ exports.deleteBook = async (req, res) => {
 }
 
 
+
+
 exports.addBook = async (req, res) => {
   try {
-    const newBook = await Book.create(req.body);
+    const newBook = await Book.create({ ...req.body, user: req.user.id });
     res.status(201).json({ message: 'Book successfully added', book: newBook });
   } catch (error) {
     console.error(error);
@@ -60,11 +85,37 @@ exports.addBook = async (req, res) => {
   }
 }
 
+
+
+
+
 exports.updateBook = async (req, res) => {
   try {
+
+    // Check if the ID is valid or not
     if (!isValidObjectId(req.params.id)) {
       return res.status(400).json({ error: 'invalid ID'});
     }
+    
+    // Find The book by the ID
+    const book = await Book.findById(req.params.id);
+    if (!book) {
+      return res.status(400).json({ error: 'Book not found'});
+    }
+
+    // Check if the user is the owner of the book
+    const user = await User.findById(req.user.id);
+    if (!book.user) {
+      return res.status(400).json({ error: 'User not found' });
+    }
+
+
+    // Verify that the user who want to update the book is the owner
+    if (book.user.toString() !== user.id) {
+      return res.status(401).json({ error: 'Only the owner can delete this book' });;
+    }
+
+    // Update the book
     const bookToUpdate = await Book.findByIdAndUpdate(req.params.id, req.body, { new: true, useFindAndModify: false });
     if (!bookToUpdate) {
       return res.status(400).json({ error: 'Book not found'});
